% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb_balance_rmosek_ate.R
\name{tfb_balance_rmosek_ate}
\alias{tfb_balance_rmosek_ate}
\title{Targeted Function Balancing Optimization with RMOSEK -- ATE}
\usage{
tfb_balance_rmosek_ate(
  X_c,
  X_t,
  beta_c,
  beta_t,
  sqrtV_c,
  sqrtV_t,
  sigma2_c,
  sigma2_t,
  i,
  d,
  chi_q,
  quiet,
  rtol
)
}
\arguments{
\item{X_c}{covariate matrix used to fit f_0}

\item{X_t}{covariate matrix used to fit f_1}

\item{beta_c}{coefficient vector for the initial regression on the control set}

\item{beta_t}{coefficient vector for the initial regression on the test set}

\item{sqrtV_c}{square root of variance matrix for the initial regression on the control set}

\item{sqrtV_t}{square root of variance matrix for the initial regression on the test set}

\item{sigma2_c}{estimated residual variance for the initial regression on the control set}

\item{sigma2_t}{estimated residual variance for the initial regression on the test set}

\item{i}{fold indices}

\item{d}{treatment status vector}

\item{chi_q}{probability threshold}

\item{quiet}{whether to suppress console output}

\item{rtol}{tolerance level for MOSEK solver}
}
\value{
A numeric vector of weights.
}
\description{
This function solves the linear optimization problem presented by TFB with MOSEK, given information from an initial regression, with the ATE as the estimand.
}
\examples{
set.seed(1221)
i <- sample(rep(c(TRUE,FALSE),75))
# only observations in the split are used to fit the initial regression
X <- as.matrix(iris[, 2:4])[i,]
d <- iris[, 5] == "setosa"
# coefficients for iris data from tfb_target_ols
beta_c <- c(0.565,0.812,-0.686)
beta_t <- c(0.532,0.632,-0.251)
# root of covariance matrix for iris data from tfb_sqrtV
sqrtV_c <- matrix(
  c(0.177,-0.017,-0.0143,
  -0.017,0.0904,-0.0614,
  -0.0143,-0.0614,0.199),nrow=3
)
sqrtV_t <- matrix(
  c(0.158,-0.0511,0.00316,
  -0.0511,0.423,-0.126,
  0.00316,-0.126,0.614),nrow=3
)
# standard error for iris data from tfb_target_ols
sigma2_c <- 0.12
sigma2_t <- 0.0616
tfb:::tfb_balance_rmosek_ate(
  X,X,beta_c,beta_t,sqrtV_c,sqrtV_t,sigma2_c,sigma2_t,i,d,0.95,TRUE, 1e-6
)
}
\keyword{tfb}
