% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb_balance_rmosek_att.R
\name{tfb_balance_rmosek_att}
\alias{tfb_balance_rmosek_att}
\title{Targeted Function Balancing Optimization with RMOSEK -- ATT}
\usage{
tfb_balance_rmosek_att(X, beta, sqrtV, sigma2, i, d, chi_q, quiet, rtol)
}
\arguments{
\item{X}{covariate matrix}

\item{beta}{coefficient vector}

\item{sqrtV}{square root of variance matrix}

\item{sigma2}{estimated variance}

\item{i}{fold indices}

\item{d}{treatment status vector}

\item{chi_q}{probability threshold}

\item{quiet}{whether to suppress console output}

\item{rtol}{tolerance level for MOSEK solver}
}
\value{
A numeric vector of weights.
}
\description{
This function solves the linear optimization problem presented by TFB with MOSEK, given information from an initial regression, with the ATT as the estimand.
}
\examples{
set.seed(1221)
i <- sample(rep(c(TRUE,FALSE),75))
# only observations in the split are used to fit the initial regression
X <- as.matrix(iris[, 2:4])[i,]
d <- iris[, 5] == "setosa"
# coefficients for iris data from tfb_target_ols
beta <- c(0.565,0.812,-0.686)
# root of covariance matrix for iris data from tfb_sqrtV
sqrtV <- matrix(
  c(0.177,-0.017,-0.0143,
  -0.017,0.0904,-0.0614,
  -0.0143,-0.0614,0.199),nrow=3
)
# standard error for iris data from tfb_target_ols
sigma2 <- 0.12
tfb:::tfb_balance_rmosek_att(X,beta,sqrtV,sigma2,i,d,0.95,TRUE, 1e-6)
}
\keyword{tfb}
