% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb_betafn_bart.R
\name{tfb_betafn_bart}
\alias{tfb_betafn_bart}
\title{BART Beta Function}
\usage{
tfb_betafn_bart(X, y, params)
}
\arguments{
\item{X}{covariate matrix}

\item{y}{outcome vector}

\item{params}{additional arguments}
}
\description{
returns a bootstrapped beta for BART
}
\examples{
tfb_betafn_bart()
}
\keyword{tfb}
