% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb_betafn_krls.R
\name{tfb_betafn_krls}
\alias{tfb_betafn_krls}
\title{KRLS Beta Function}
\usage{
tfb_betafn_krls(X, y, params)
}
\arguments{
\item{X}{covariate matrix}

\item{y}{outcome vector}

\item{params}{additional arguments}
}
\value{
A numeric vector of coefficients.
}
\description{
This is a helper function that returns a bootstrapped beta for KRLS.
}
\examples{
set.seed(1221)
i <- sample(rep(c(T,F),75)) # fold indices for iris data from tfb_fold
d <- iris[, 5] == "setosa"
X <- as.matrix(iris[!i & !d, 2:4])
y <- iris[!i & !d, 1]
boot <- sample(1:nrow(X), replace=T)
tfb:::tfb_betafn_krls(X[boot,],y[boot],list())
}
\keyword{tfb}
