% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb_bootstrapped_covariance.R
\name{tfb_bootstrapped_covariance}
\alias{tfb_bootstrapped_covariance}
\title{Bootstrapping The Covariance Matrix}
\usage{
tfb_bootstrapped_covariance(X, y, betafn, bstrap_reps, params)
}
\arguments{
\item{X}{covariates}

\item{y}{responses}

\item{betafn}{function taking X,y that returns beta}

\item{bstrap_reps}{number of repetitions}

\item{params}{additional arguments for the fit}
}
\value{
A numeric matrix, the bootstrapped covariance matrix of the model.
}
\description{
This function bootstraps a given fit's covariance matrix by re-sampling with replacement and re-fitting the model repeatedly.
}
\examples{
set.seed(1221)
# fold indices for iris data from tfb_fold
i <- sample(rep(c(TRUE,FALSE),75))
d <- iris[, 5] == "setosa"
X <- as.matrix(iris[!i & !d, 2:4])
y <- iris[!i & !d, 1]
tfb:::tfb_bootstrapped_covariance(X,y,tfb:::tfb_betafn_elasticnet,100,list())
}
\keyword{tfb}
