% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb_fold.R
\name{tfb_fold}
\alias{tfb_fold}
\title{Sample Splitting}
\usage{
tfb_fold(d, folds)
}
\arguments{
\item{d}{treatment status vector}

\item{folds}{number of folds to split the data across}
}
\value{
A list of logical vectors. The ith vector specifies which observations belong to the ith fold.
}
\description{
This function splits a covariate matrix into folds for TFB.
}
\examples{
d <- iris[, 5] == "setosa"
set.seed(1221)
tfb:::tfb_fold(d,2)
}
\keyword{tfb}
