% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb_gram_matrix.R
\name{tfb_gram_matrix}
\alias{tfb_gram_matrix}
\title{Kernel Matrix for TFB with KRLS}
\usage{
tfb_gram_matrix(X_rows, X_cols, kernelfn)
}
\arguments{
\item{X_rows}{covariate matrix defining rows of K}

\item{X_cols}{covariate matrix defining cols of K}

\item{kernelfn}{inner product to use}
}
\value{
A numeric matrix, the kernel/gram matrix produced by \code{X_rows} and \code{X_cols}.
}
\description{
This function creates a kernel matrix with size n_rows by n_cols using an inner product. for each row of \code{X_rows}, its inner product with (similarity to) every row of \code{X_cols} is calculated.
}
\examples{
X <- as.matrix(iris[,2:4])
d <- iris[, 5] == "setosa"
set.seed(1221)
# fold indices for iris data from tfb_fold
i <- sample(rep(c(TRUE,FALSE),75))
tfb:::tfb_gram_matrix(X[i, ], X[!i & d, ], tfb:::tfb_kernelfn)
}
\keyword{tfb}
