% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb_kernelfn.R
\name{tfb_kernelfn}
\alias{tfb_kernelfn}
\title{Default Kernel Function}
\usage{
tfb_kernelfn(X_i, X_j)
}
\arguments{
\item{X_i}{ith row of X}

\item{X_j}{jth row of X}
}
\value{
A double, the guassian kernel of \code{X_i} and \code{X_j}.
}
\description{
This is an inner product function, taking two rows of X and returning a value. The default is the gaussian kernel with Hainmueller and Hazlett's suggested bandwidth parameter of \code{ncol(X)}. However, if a non-default \code{kernel} parameter is used when fitting KRLS, the user will have to supply the equivalent kernel function in order for \code{tfb} to return correct estimates. This is because KRLS standardizes its matrices for simpler computation, so in order to cross-fit KRLS \code{tfb} needs to be able to re-standardize the kernel matrix used in optimization.
}
\examples{
X <- as.matrix(iris[,2:4])
tfb:::tfb_kernelfn(X[1,],X[2,])
}
\keyword{tfb}
