% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_sampledat2.R
\docType{data}
\name{tfb_sampledat2}
\alias{tfb_sampledat2}
\title{TFB Sample Dataset 2}
\format{
\subsection{\code{tfb_sampledat2}}{

Well suited for a nonlinear fit.
A numeric matrix with 1,000 rows and 6 columns:
\describe{
\item{x1,x2}{covariates}
\item{d}{treatment}
\item{y}{observed outcomes}
\item{y0,y1}{potential outcomes}
}
}
}
\usage{
tfb_sampledat2
}
\description{
This is a sample dataset for use with \code{tfb}.
The dataset is theoretical, so the covariates, treatments, and outcomes do not have any special meaning.
More data of the same distribution can be generated using the dgp found in data-raw/.
The treatment assignment mechanism is known for this dataset, which makes it useful for comparing the efficiency of causal inference estimators.
}
\keyword{datasets}
