% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb_sqrtV.R
\name{tfb_sqrtV}
\alias{tfb_sqrtV}
\title{TFB Covariance Matrix Square Root}
\usage{
tfb_sqrtV(V, ev_approx, ev_n, quiet, sample, fold, treatment)
}
\arguments{
\item{V}{covariance matrix}

\item{ev_approx}{whether to approximate the eigenpairs}

\item{ev_n}{how many eigenvalues to approximate}

\item{quiet}{whether to suppress console output}

\item{sample}{sample of the current iteration}

\item{fold}{fold of the current iteration}

\item{treatment}{target treatment value of the current iteration}
}
\value{
A numeric matrix, the square root of \code{V}.
}
\description{
This function returns the square root of the covariance matrix or an approximation.
}
\examples{
# covariance matrix for iris data from tfb_target_ols
V <- matrix(
  c(0.0319,-0.00368,-0.00433,
  -0.00368,0.0122,-0.0175,
  -0.00433,-0.0175,0.0436),nrow=3
)
tfb:::tfb_sqrtV(V,FALSE,NULL,TRUE,1,1,"c")
}
\keyword{tfb}
