% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb_summary.R
\name{tfb_summary}
\alias{tfb_summary}
\title{Final Summary Statistics for TFB}
\usage{
tfb_summary(d, y, estimand, confidence, wdims_samples, v_hats_samples)
}
\arguments{
\item{d}{treatment vector}

\item{y}{outcome vector}

\item{estimand}{causal inference estimand being estimated.}

\item{confidence}{confidence level for the CI produced by tfb}

\item{wdims_samples}{vector of wdims, one for each sample}

\item{v_hats_samples}{vector of variances, one for each sample}
}
\value{
A list:
\itemize{
\item \code{estimate}: the final tfb estimate
\item \code{variance}: the final variance of the tfb estimate
\item \code{confidence_interval}: the confidence interval for the estimate
\item \code{p_value}: the two-sided p-value for the estimate
\item \code{dim}: the naive difference in means for the data
}
}
\description{
This function produces the final summary statistics for TFB across all samples.
}
\examples{
d <- iris[, 5] == "setosa"
y <- iris[, 1]
wdims_samples <- -0.0315 # wdim for the iris data from tfb_summary_sample
v_hats_samples <- 0.0278 # variance for the estimate for the iris data from tfb_summary_sample
tfb:::tfb_summary(d,y,"att",0.95,wdims_samples,v_hats_samples)
}
\keyword{tfb}
