% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfb_target_bart.R
\name{tfb_target_bart}
\alias{tfb_target_bart}
\title{TFB BART Initial Regression}
\usage{
tfb_target_bart(
  X,
  d,
  y,
  i,
  bstrap_cov,
  bstrap_reps,
  reg_d,
  treatment,
  quiet,
  params
)
}
\arguments{
\item{X}{covariate matrix}

\item{d}{treatment vector}

\item{y}{outcome vector}

\item{i}{split indices}

\item{bstrap_cov}{whether to bootstrap covariance}

\item{bstrap_reps}{how many times to bootstrap covariance}

\item{reg_d}{whether to include the treatment as a variable for regression}

\item{treatment}{target treatment value}

\item{params}{additional named parameters to the fit}
}
\description{
Returns initial regression information for BART in a specific fold of the data
}
\examples{
tfb_target_bart()
}
\keyword{tfb}
